+++ src/models/context_tag_keys.rs
@@ -31,6 +31,11 @@ impl Tags {
     pub(crate) fn get(&self, key: &ContextTagKey) -> Option<&String> {
         self.0.get(key.key)
     }
+
+    #[cfg(feature = "metrics")]
+    pub(crate) fn is_empty(&self) -> bool {
+        self.0.is_empty()
+    }
 }
 
 macro_rules! context_tag_keys {
@@ -183,7 +188,7 @@ context_tag_keys! {
     CLOUD_ROLE_INSTANCE("ai.cloud.roleInstance", 256),
 
     /// SDK version. See
-    /// https://github.com/Microsoft/ApplicationInsights-Home/blob/master/SDK-AUTHORING.md#sdk-version-specification
+    /// <https://github.com/Microsoft/ApplicationInsights-Home/blob/master/SDK-AUTHORING.md#sdk-version-specification>
     /// for information.
     INTERNAL_SDK_VERSION("ai.internal.sdkVersion", 64),